// src/api/model/exam.ts
/**
 * Generated by orval v6.26.0 🍺
 * Do not edit manually.
 * student-hub-api
 * Api Documentation for StudentHub API
 * OpenAPI spec version: 0.0
 */
import type { ExerciseBasicInfo } from './exerciseBasicInfo';

export interface Exam {
  bestGrade: number;
  bestSubTopic: string;
  criteriaPdf: string;
  examId: number;
  examName: string;
  examPdf: string;
  exercises: ExerciseBasicInfo[];
  isDone: boolean;
  solutionPdf: string;
  worstSubTopic: string;
}